requirements.txt
auto-ml
awscli
beautifulsoup4
coreapi
cssselect
cufflinks
dj-database-url
Django==2.0.6
django-crispy-forms
django-debug-toolbar
django-sslserver
djangorestframework
Flask
gensim
gevent
google-api-python-client
graphviz
greenlet
Jinja2
matplotlib
numpy
opencv-python
pandas
pandas-datareader
Pillow
plotly
promise
pytest
pytest-cov
python-dateutil
python-memcached
python-utils
requests
requests-file
requests-ftp
requests-kerberos
requests-oauthlib
scapy
scikit-image
scikit-learn
scikit-MDR
scikit-plot
scikit-surprise
scipy
Scrapy
seaborn
sklearn-pandas
spacy
spark-sklearn
sparkit-learn
SQLAlchemy
sqlparse
stanfordcorenlp
statsmodels
sumy
sympy
TPOT
yellowbrick
